add breadcrumb in edit page and change test
